Add support for WIP blog entries
Markdown files under src/blog/ beginning with 'wip' will be built and
placed in the out/ directory, but will *not* be given an entry in the
auto generated blog.html page.
This allows work-in-progress entries to be safely committed and pushed
to the server without having to be "officially" published.