[BLOG] Remove "Writing and Interpreter in Go in Python"
Turns out I didn't really have a lot of useful things to say. I really
liked the book, but other than me saying "I enjoyed working through this
book and writing my interpreter" and "I wrote my interpreter in Python
because I am not a fan of Go (just look at the amount of boilerplate in
the book)" there was not much I could really add that was worth reading.